refactor: 단일책임원칙에 따른 메서드 분리 및 클래스 재설계
- Delimiter 클래스 유틸성 클래스로 분리
- StringCalculator에 유의미한 연산 기능 추가
