+++ src/main/java/stringcalculator/StringCalculatorMain.java
@@ -1,6 +1,6 @@
 package stringcalculator;
 
-import stringcalculator.domain.Delimiter;
+import stringcalculator.util.Delimiter;
 import stringcalculator.domain.StringCalculator;
 import stringcalculator.util.Converter;
 import stringcalculator.view.InputView;
@@ -11,9 +11,7 @@ public class StringCalculatorMain {
 
     public static void main(String[] args) {
         String input = InputView.getExpression();
-        Delimiter delimiter = new Delimiter(input);
-        List<Integer> numbers = Converter.convertOperandType(delimiter.getNumbers());
-        int result = StringCalculator.execute(numbers);
+        int result = StringCalculator.add(input);
         ResultView.printResult(result);
     }
 }
